feat(ZH-99): API 호출 로직 분리
